+++ src/app/core/helpers/favoriteCoin.helper.ts
